- Fix issue of default limits fetched from dbcc. - Change encoding logic.
E.g.: Fill up signal to zero in case of out of range value while encoding,
Do modification in this area to avoid unnecessary zero value in real signal even for the very short time.
This was creating problem when real signal encoded with zero in healthy condition and task switched to CAN transmit.

Signed-off-by: Kalpesh Ramani <kalpeshramani498@gmail.com>